Load table data in an event.
The data retrieval function is now built by a call to mk_revision_search_func prior to
calling mk_table. An event listener is added that when triggered, fetches the revision
data and then calls fill_data with the results.

fill_data is also optimized by placing the data row elements in an array, then using
.replaceChildren() to display the content.